add support for nano33ble using mbed arduino core (#53)
The Arduino core for the nano33ble uses mbed and wraps it in the Arduino
core API. Therefore __MBED__ is defined, which makes JLed think it runs
in an Mbed environment. But in this case, JLed needs to use the Arduino
API. The problem is solved by checking also for ARDUINO_API_VERSION:
  #if defined(__MBED__) && !defined(ARDUINO_API_VERSION)
  -> running in mbed environment
in jled.h.

See
 * https://github.com/arduino/ArduinoCore-API
 * https://github.com/arduino/ArduinoCore-nRF528x-mbedos